// templateCopy/GoodSleepEIP/ClientApp/src/service/apiServices.schemas.ts
export interface Permissions {
  PermissionId: number;
  /** @nullable */
  ModuleName: string | null;
  /** @nullable */
  PermissionModuleSubType: string | null;
  /** @nullable */
  PermissionDescription?: string | null;
  IsPublic: boolean;
  CanRead: boolean;
  CanReadAll: boolean;
  CanCreate: boolean;
  CanUpdate: boolean;
  CanDelete: boolean;
  CanManage: boolean;
  IsSystemReserved: boolean;
  CreationTime: string;
}
// ...
export interface UserSessionData {
  UserId: number;
  /** @nullable */
  Username: string | null;
  /** @nullable */
  UserDescription: string | null;
  DepartmentId?: number;
  /** @nullable */
  UserEmail?: string | null;
  /** @nullable */
  LineUserId?: string | null;
  /** @nullable */
  Token?: string | null;
  /** @nullable */
  Permissions?: Permissions[] | null;
  /** @nullable */
  RefreshToken?: string | null;
  /** @nullable */
  RefreshTokenExpiry?: string | null;
  /** @nullable */
  PersonId?: string | null;
  /** @nullable */
  PersonName?: string | null;
}